refactor: split link toolbar into input and utils modules
